Confirm plannings delete operation

+++ src/pages/main-plannings/plannings.ts
@@ -6,6 +6,9 @@ import { DiaPlanningsService } from '../../services/dia-plannings-service';
 import { Planning } from '../../models/plannings-model';
 import { ModalController } from 'ionic-angular/components/modal/modal-controller';
 import { PlanningsEditorPage } from '../plannings-editor/plannings-editor';
+import { DiaMessage } from '../../models/messages-model';
+import { forkJoin } from 'rxjs/observable/forkJoin';
+import { DiaMessageService } from '../../services/dia-message-service';
 
 @Component({
   selector: 'tab-plannings',
@@ -18,6 +21,7 @@ export class PlanningsPage {
   constructor(private navCtrl: NavController,
               private navParams: NavParams,
               private translate: TranslateService,
+              private messageService: DiaMessageService,
               private planningsService: DiaPlanningsService,
               private modalCtrl: ModalController) {
 
@@ -48,14 +52,25 @@ export class PlanningsPage {
   }
 
   deletePlanning(planning) {
-    this.planningsService.deletePlanning(planning.id).subscribe(
-      (resp) => {
-        this.refresh();
-      },
-      (err) => {
-        console.log(err);
-      }
-    )
+    forkJoin(
+      this.translate.get("Delete Planning?"),
+      this.translate.get("Are you sure to delete this planning?")
+    ).subscribe(([title, message]) => {
+      let diamessage = new DiaMessage(title, "info", message)
+      this.messageService.confirmMessage(diamessage).subscribe((ok) => {
+        // if we have confirmation, delete the planning
+        if (ok) {
+          this.planningsService.deletePlanning(planning.id).subscribe(
+            (resp) => {
+              this.refresh();
+            },
+            (err) => {
+              console.log(err);
+            }
+          );
+        }
+      });
+    });
   }
 
   refresh() {
